# docs/api/openapi.json
{
  "components": {
    "schemas": {
      "AnalysisCreateRequest": {
        "description": "Request body for creating a new analysis run.",
        "properties": {
          "text": {
            "minLength": 1,
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "text"
        ],
        "title": "AnalysisCreateRequest",
        "type": "object"
      },
      "AnalysisRunResponse": {
        "description": "API response schema for a persisted analysis run.",
        "properties": {
          "analysis_json": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Analysis Json"
          },
          "correlation_id": {
            "title": "Correlation Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "detected_language": {
            "title": "Detected Language",
            "type": "string"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "error_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Reason"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "input_text": {
            "title": "Input Text",
            "type": "string"
          },
          "language_confidence": {
            "title": "Language Confidence",
            "type": "number"
          },
          "model_id": {
            "title": "Model Id",
            "type": "string"
          },
          "prompt_version": {
            "title": "Prompt Version",
            "type": "string"
          },
          "run_status": {
            "title": "Run Status",
            "type": "string"
          },
          "validation_report": {
            "additionalProperties": true,
            "title": "Validation Report",
            "type": "object"
          },
          "validation_status": {
            "title": "Validation Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "correlation_id",
          "input_text",
          "analysis_json",
          "validation_report",
          "detected_language",
          "language_confidence",
          "model_id",
          "prompt_version",
          "run_status",
          "validation_status",
          "error_code",
          "error_reason",
          "created_at"
        ],
        "title": "AnalysisRunResponse",
        "type": "object"
      },
      "ErrorPayload": {
        "description": "Public API error payload.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "correlation_id": {
            "title": "Correlation Id",
            "type": "string"
          },
          "details": {
            "additionalProperties": true,
            "title": "Details",
            "type": "object"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "details",
          "correlation_id"
        ],
        "title": "ErrorPayload",
        "type": "object"
      },
      "ErrorResponse": {
        "description": "Public API error response envelope.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorPayload"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Social Cleanup Machine API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/analyses": {
      "get": {
        "operationId": "get_analyses_api_v1_analyses_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AnalysisRunResponse"
                  },
                  "title": "Response Get Analyses Api V1 Analyses Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Analyses"
      },
      "post": {
        "operationId": "create_analysis_api_v1_analyses_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalysisCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisRunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request"
          }
        },
        "summary": "Create Analysis"
      }
    },
    "/api/v1/analyses/{analysis_id}": {
      "get": {
        "operationId": "get_analysis_api_v1_analyses__analysis_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "analysis_id",
            "required": true,
            "schema": {
              "title": "Analysis Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisRunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Analysis run not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request"
          }
        },
        "summary": "Get Analysis"
      }
    },
    "/api/v1/analyses/{analysis_id}/rerun": {
      "post": {
        "operationId": "rerun_existing_analysis_api_v1_analyses__analysis_id__rerun_post",
        "parameters": [
          {
            "in": "path",
            "name": "analysis_id",
            "required": true,
            "schema": {
              "title": "Analysis Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisRunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Analysis run not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request"
          }
        },
        "summary": "Rerun Existing Analysis"
      }
    },
    "/health": {
      "get": {
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Health Health Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Health"
      }
    }
  }
}